refactor: move configuration check logic
